requirements.txt
discord.py
loguru
requests
asyncio